requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-Mail==0.9.1
python-dotenv==1.0.0
gunicorn==21.2.0
openai==1.12.0
pymongo==4.5.0
pytz==2023.3
httpx==0.25.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.1.0
google-api-python-client==2.101.0
google-auth==2.25.2